rpc: Move verifier parse_deposit_params to parsers.
