+++ core/src/rpc/parsers/verifier.rs
@@ -1,34 +1,41 @@
-use crate::rpc::clementine::DepositParams;
-use crate::EVMAddress;
+use crate::{rpc::clementine, EVMAddress};
 use bitcoin::address::NetworkUnchecked;
 use tonic::Status;
 
-pub mod verifier;
-
-pub fn parse_deposit_params(
-    deposit_params: DepositParams,
+pub fn get_deposit_params(
+    deposit_sign_session: clementine::DepositSignSession,
+    verifier_idx: usize,
 ) -> Result<
     (
         bitcoin::OutPoint,
         EVMAddress,
         bitcoin::Address<NetworkUnchecked>,
         u16,
+        u32,
     ),
     Status,
 > {
+    let deposit_params = deposit_sign_session
+        .deposit_params
+        .ok_or(Status::invalid_argument("No deposit outpoint received"))?;
+
     let deposit_outpoint: bitcoin::OutPoint = deposit_params
         .deposit_outpoint
         .ok_or(Status::invalid_argument("No deposit outpoint received"))?
         .try_into()?;
-    let evm_address: EVMAddress = deposit_params
-        .evm_address
-        .try_into()
-        .map_err(|_| Status::invalid_argument("Could not parse deposit outpoint EVM address"))?;
+    let evm_address: EVMAddress = deposit_params.evm_address.try_into().map_err(|e| {
+        Status::invalid_argument(format!(
+            "Failed to convert evm_address to EVMAddress: {}",
+            e
+        ))
+    })?;
     let recovery_taproot_address = deposit_params
         .recovery_taproot_address
         .parse::<bitcoin::Address<_>>()
         .map_err(|e| Status::internal(e.to_string()))?;
     let user_takes_after = deposit_params.user_takes_after;
+    let session_id = deposit_sign_session.nonce_gen_first_responses[verifier_idx].id;
+
     Ok((
         deposit_outpoint,
         evm_address,
@@ -39,5 +46,6 @@ pub fn parse_deposit_params(
                 e
             ))
         })?,
+        session_id,
     ))
 }
